refactor: re-organize `app` directory structure
